digest times in GMT, not local timezone
Although group activity appears in the correct timezone in regular
buddypress pages, the buddypress group activity daily/weekly digest was
reporting all the activiity in GMT, not the timezone we set in
wordpress.
It turned out the fix was just a one-liner. Somehow the timezone wasn't
getting set, so I added this line to set it to the timezone set in
wordpress.