Add bbPress 2.2 support to the 'New Topics' subscription option.
If group status is 'sub', this means the user is subscribed to 'New Topics'.

However, 'New Topics' was refactored awhile ago to be recorded in
ass_group_notification_forum_posts() for the bundled forums.

So, now we can use this to our advantage and support bbPress 2 by checking
the 'bbp_topic_create' activity type.