Block incomplete bbPress activity entries from being recorded in GES.
When bbPress first records an item into the group activity stream, it is
incomplete as it is first recorded on the 'wp_insert_post' action.

bbPress later updates the activity item on the 'bbp_new_reply' /
'bbp_new_topic' action.

To workaround this, we do some checks in
ass_default_block_group_activity_types() to block the incomplete activity
items from being recorded or sent by GES.

See cuny-academic-commons/commons-in-a-box#51.