Reinstate bbPress' topic subscription in BuddyPress groups.
If the current user is not subscribed to the group's 'All Mail' option,
allow the user to use bbPress' built-in topic subscription functionality.

This emulates GES' old 'Follow / Mute' feature.  (Note: This is managed
by bbPress and *not* by GES.)

If the current user is subscribed to the group's 'All Mail' option, we
disable this bbP's functionality to prevent duplicate emails.