Fix notice generated by bbPress when a forum post is submitted.
bbPress 2.5.4 changed the way emails are sent out and broke a filter that
GES was using.  bbP 2.5.4 now sends out one email to topic subscribers and
BCCs everyone that is subscribed to that topic.

For bbPress 2.5.4, this commit filters the list of topic subscribers and
removes the subscriber from the list if the subscriber is already
subscribed to the group's "All Mail" option in GES.  This is done to
prevent duplicate emails.

Fixes #59.