Add asynchronous handling of the digest mailing process.
Instead of trying to send all of the digests and summaries in a single
page load, this moves the process off to a `WP_Async_Task` hook that
starts the process on a fresh page load. It also breaks the mailing
down into smaller chunks, so that we’re only trying to send 50
(filterable number) digests before spawning a new page load to send the
next batch.