Fix changing member email status on a group's "Manage > Members" for BP 2.7.
BP 2.7 added a parameter to the 'bp_group_manage_members_admin_item' hook.

This broke functionality in detecting the group member user ID in the
group member loop.

Commit fixes this by checking if the $user_id parameter is not numeric.
If not numeric, we fallback to the group member user ID as before.