Better digest handling of deleted activity items.
When an activity item is recorded in a digest, and the activity
item is deleted before the digest is sent, the digest assembler will
not have any source data to work with. When this happens for *all*
activity items associated with a specific group (within a digest period),
that group's section should be omitted from the digest. And when
all of a user's digest group's are empty, the user's digest should
not be sent for that day.