Require a form submission when clicking on the "Unsubscribe" link from the email footer.
Previously, if a user clicked on the "Unsubscribe" link, it would
unsubscribe the user from the group (or all groups depending on the
setting) without any prior warning.  This can be a problem if email clients
use some form of link auto-scanning.

This commit changes this behavior so a user form submission is required to
confirm the group email unsubscription.