Changes to the way the PHPMailer object is filtered during digest sends.
This is broken into a separate commit because it likely is a functional
change in addition to one designed to placate PHPCS. The callbacks
used to filter the from_name and from_email props of PHPMailer were
not properly set up to use the out-of-scope variables. I've fixed
that bug here, but it may result in new behavior.